Add files via upload
This is the repository for the project we'll build together during TylerMcGinnis.com's "React Fundamentals" course, for more information on the course see tylermcginnis.com/courses
You can see the final project here  https://rt-react-fundamentals.firebaseapp.com/
This is a "Github Battle App". Enter in two Github usernames, and it'll declare a winner.